% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add documents from R objects}
\usage{
add(x, conn, commit = TRUE, commit_within = NULL, overwrite = TRUE,
  boost = NULL, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{x}{Documents, either as rows in a data.frame, or a list.}

\item{conn}{Connection object. Required. See \code{\link{solr_connect}}.}

\item{commit}{(logical) If \code{TRUE}, documents immediately searchable.
Default: \code{TRUE}}

\item{commit_within}{(numeric) Milliseconds to commit the change, the document will be added
within that time. Default: NULL}

\item{overwrite}{(logical) Overwrite documents with matching keys. Default: \code{TRUE}}

\item{boost}{(numeric) Boost factor. Default: NULL}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to
parse}

\item{raw}{(logical) If \code{TRUE}, returns raw data in format specified by
\code{wt} param}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Add documents from R objects
}
\examples{
\dontrun{
conn <- solr_connect()

# Documents in a list
ss <- list(list(id = 1, price = 100), list(id = 2, price = 500))
add(x = ss, conn)

# Documents in a data.frame
## Simple example
df <- data.frame(id = c(67, 68), price = c(1000, 500000000))
add(x = df, conn)

## More complex example, get file from package examples
file <- system.file("examples", "books.csv", package = "solr")
x <- read.csv(file, stringsAsFactors = FALSE)
class(x)
head(x)
add(x, conn)

# Use modifiers
add(x, conn, commit_within = 5000)
}
}

