% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{collapse_pivot_names}
\alias{collapse_pivot_names}
\title{Collapse Pivot Field and Value Columns}
\usage{
collapse_pivot_names(data)
}
\arguments{
\item{data}{a \code{data.frame} with every 2 columns
representing a field and value and the final representing
a count}
}
\value{
a \code{data.frame}
}
\description{
Convert a table consisting of columns in sets of 3
into 2 columns assuming that the first column of every set of 3
(field) is duplicated throughout all rows and should be removed.
This type of structure is usually returned by facet.pivot responses.
}
\keyword{internal}

