% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{collectargs}
\alias{collectargs}
\title{Function to make a list of args passing arg names through multiargs function.}
\usage{
collectargs(x)
}
\arguments{
\item{x}{Value}
}
\description{
Function to make a list of args passing arg names through multiargs function.
}

