% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_createshard.R
\name{collection_createshard}
\alias{collection_createshard}
\title{Create a shard}
\usage{
collection_createshard(name, shard, createNodeSet = NULL, raw = FALSE, ...)
}
\arguments{
\item{name}{(character) Required. The name of the collection that includes the shard
that will be splitted.}

\item{shard}{(character) Required. The name of the shard to be created.}

\item{createNodeSet}{(character) Allows defining the nodes to spread the new
collection across. If not provided, the CREATE operation will create shard-replica
spread across all live Solr nodes. The format is a comma-separated list of
node_names, such as localhost:8983_solr, localhost:8984_s olr, localhost:8985_solr.}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Create a shard
}
\examples{
\dontrun{
solr_connect()
## FIXME - doesn't work right now
# collection_create(name = "trees")
# collection_createshard(name = "trees", shard = "newshard")
}
}

