% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collection_delete.R
\name{collection_delete}
\alias{collection_delete}
\title{Add a collection}
\usage{
collection_delete(conn, name, verbose = TRUE, raw = FALSE, ...)
}
\arguments{
\item{conn}{Connection object. Required. See \code{\link{solr_connect}}.}

\item{name}{The name of the collection to be created. Required}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{raw}{(logical) If \code{TRUE}, returns raw data in format specified by
\code{wt} param}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Add a collection
}
\examples{
\dontrun{
conn <- solr_connect()
collection_create(conn, name = "helloWorld")
collection_delete(conn, name = "helloWorld")
}
}

