% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_deletealias.R
\name{collection_deletealias}
\alias{collection_deletealias}
\title{Delete a collection alias}
\usage{
collection_deletealias(conn, alias, raw = FALSE, ...)
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{alias}{(character) Required. The alias name to be created}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Delete a collection alias
}
\examples{
\dontrun{
(conn <- SolrClient$new())
conn$collection_create(name = "thingsstuff", numShards = 2)
conn$collection_createalias("tstuff", "thingsstuff")
conn$collection_clusterstatus()$cluster$collections$thingsstuff$aliases # new alias
conn$collection_deletealias("tstuff")
conn$collection_clusterstatus()$cluster$collections$thingsstuff$aliases # gone
}
}
