% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\alias{cores}
\title{List collections or cores}
\usage{
collections(conn, ...)

cores(conn, ...)
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
character vector
}
\description{
List collections or cores
}
\details{
Calls \code{\link[=collection_list]{collection_list()}} or \code{\link[=core_status]{core_status()}} internally,
and parses out names for you.
}
\examples{
\dontrun{
# connect
(conn <- SolrClient$new())

# list collections
conn$collection_list()
collections(conn)

# list cores
conn$core_status()
cores(conn)
}
}
