% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_requeststatus.R
\name{core_requeststatus}
\alias{core_requeststatus}
\title{Request status of asynchronous CoreAdmin API call}
\usage{
core_requeststatus(conn, requestid, raw = FALSE, callopts = list())
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{requestid}{The name of one of the cores to be removed. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{callopts}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Request status of asynchronous CoreAdmin API call
}
\examples{
\dontrun{
# start Solr with Schemaless mode via the schemaless eg:
#   bin/solr start -e schemaless

# FIXME: not tested yet...
# (conn <- SolrClient$new())
# conn$core_requeststatus(requestid = 1)
}
}
