% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_json.R
\name{create_json}
\alias{create_json}
\title{Create documents}
\usage{
create_json(base = "http://localhost:8983", files, wt = "json",
  verbose = TRUE, raw = FALSE, callopts = list())
}
\arguments{
\item{base}{(character) URL endpoint. This is different from the other functions in that we aren't
hitting a search endpoint. Pass in here}

\item{files}{Path to file to load into Solr}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to parse}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{raw}{(logical) If TRUE, returns raw data in format specified by wt param}

\item{callopts}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Create documents
}
\details{
You likely may not be able to run this function against many public Solr
services, but should work locally.
}
\examples{
\dontrun{
# from files
## from json, json is the default wt value
create(files = "~/books.json")
## from xml
create(files = "~/hd.xml", wt="xml")
}
}
\seealso{
Other create: \code{\link{create_xml}}
}

