% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.sr_facet}
\alias{is.sr_facet}
\alias{is.sr_high}
\alias{is.sr_search}
\title{Test for sr_facet class}
\usage{
is.sr_facet(x)

is.sr_high(x)

is.sr_search(x)
}
\arguments{
  \item{x}{Input}
}
\description{
Test for sr_facet class

Test for sr_high class

Test for sr_search class
}

