% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\title{Get the schema for a collection or core}
\usage{
schema(conn, name = NULL, wt = "json", raw = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{conn}{Connection object. Required. See \code{\link{solr_connect}}.}

\item{name}{(character) Name of collection or core}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to
parse}

\item{raw}{(logical) If \code{TRUE}, returns raw data in format specified by
\code{wt} param}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get the schema for a collection or core
}
\examples{
\dontrun{
conn <- solr_connect()

# start Solr in Schemaless mode: bin/solr start -e schemaless
# schema(conn, "gettingstarted")

# start Solr in Standalone mode: bin/solr start
# then add a core: bin/solr create -c helloWorld
# schema(conn, "helloWorld")
}
}

