\docType{package}
\name{solr}
\alias{solr}
\alias{solr-package}
\title{solr - a general purpose package to interact with Solr endpoints}
\description{
  The solr package is an R interface to Solr. Currently (as
  of 2013-12-07) this package only does the getting data
  part, not writing data, but if you want data writing
  capability do speak up and/or send a pull request.
}
\details{
  There are currently only three functions:

  \describe{ \item{solr_search}{General search}
  \item{solr_facet}{Faceting}
  \item{solr_highlight}{Highlighting} }
}

