% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add documents from R objects}
\usage{
add(x, commit = TRUE, wt = "json", raw = FALSE,
  base = "http://localhost:8983", ...)
}
\arguments{
\item{x}{Documents, either as rows in a data.frame, or a list.}

\item{commit}{(logical) If \code{TRUE}, documents immediately searchable.
Deafult: \code{TRUE}}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to
parse}

\item{raw}{(logical) If \code{TRUE}, returns raw data in format specified by
\code{wt} param}

\item{base}{(character) URL endpoint. This is different from the other functions in
that we aren't hitting a search endpoint. Pass in here}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Add documents from R objects
}
\examples{
\dontrun{
# Documents in a list
ss <- list(list(id = 1, price = 100), list(id = 2, price = 500))
add(x = ss)

# Documents in a data.frame
df <- data.frame(id = c(67, 68), price = c(1000, 500000000))
add(x = df)
}
}

