% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_delete.R
\name{collection_delete}
\alias{collection_delete}
\title{Add a collection}
\usage{
collection_delete(name, raw = FALSE, ...)
}
\arguments{
\item{name}{The name of the collection to be created. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Add a collection
}
\examples{
\dontrun{
solr_connect()
collection_create(name = "helloWorld")
collection_delete(name = "helloWorld")
}
}

