% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_deletealias.R
\name{collection_deletealias}
\alias{collection_deletealias}
\title{Delete a collection alias}
\usage{
collection_deletealias(alias, raw = FALSE, ...)
}
\arguments{
\item{alias}{(character) Required. The alias name to be created}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a collection alias
}
\examples{
\dontrun{
solr_connect()
collection_create(name = "thingsstuff", numShards = 2)
collection_createalias("tstuff", "thingsstuff")
collection_clusterstatus()$cluster$collections$thingsstuff$aliases # new alias
collection_deletealias("tstuff")
collection_clusterstatus()$cluster$collections$thingsstuff$aliases # gone
}
}

