% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_list.R
\name{collection_list}
\alias{collection_list}
\title{List collections}
\usage{
collection_list(raw = FALSE, ...)
}
\arguments{
\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
List collections
}
\examples{
\dontrun{
solr_connect()
collection_list()
collection_list()$collections
}
}

