% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_list.R
\name{collection_list}
\alias{collection_list}
\title{List collections}
\usage{
collection_list(conn, raw = FALSE, ...)
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{raw}{(logical) If \code{TRUE}, returns raw data in format specified by
\code{wt} param}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
List collections
}
\examples{
\dontrun{
(conn <- SolrClient$new())

conn$collection_list()
conn$collection_list()$collections
collection_list(conn)
}
}
