% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_overseerstatus.R
\name{collection_overseerstatus}
\alias{collection_overseerstatus}
\title{Get overseer status}
\usage{
collection_overseerstatus(raw = FALSE, ...)
}
\arguments{
\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Returns the current status of the overseer, performance statistics
of various overseer APIs as well as last 10 failures per operation type.
}
\examples{
\dontrun{
solr_connect()
collection_overseerstatus()
res <- collection_overseerstatus()
res$responseHeader
res$leader
res$overseer_queue_size
res$overseer_work_queue_size
res$overseer_operations
res$collection_operations
res$overseer_queue
res$overseer_internal_queue
res$collection_queue
}
}

