% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_reload.R
\name{collection_reload}
\alias{collection_reload}
\title{Reload a collection}
\usage{
collection_reload(conn, name, raw = FALSE, ...)
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{name}{(character) The name of the core to be created. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{You can pass in parameters like \code{property.name=value}    to set
core property name to value. See the section Defining core.properties for
details on supported properties and values.
(https://lucene.apache.org/solr/guide/7_0/defining-core-properties.html)}
}
\description{
Reload a collection
}
\examples{
\dontrun{
(conn <- SolrClient$new())
conn$collection_create(name = "helloWorld")
conn$collection_reload(name = "helloWorld")
}
}
