% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_reload.R
\name{collection_reload}
\alias{collection_reload}
\title{Reload a collection}
\usage{
collection_reload(conn, name, raw = FALSE, callopts)
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{name}{(character) The name of the core to be created. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{callopts}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Reload a collection
}
\examples{
\dontrun{
(conn <- SolrClient$new())

if (!conn$collection_exists("helloWorld")) {
  conn$collection_create(name = "helloWorld")
}

conn$collection_reload(name = "helloWorld")
}
}
