% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_reload.R
\name{collection_reload}
\alias{collection_reload}
\title{Reload a collection}
\usage{
collection_reload(name, raw = FALSE, ...)
}
\arguments{
\item{name}{The name of the collection to reload. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Reload a collection
}
\examples{
\dontrun{
solr_connect()
collection_create(name = "helloWorld")
collection_reload(name = "helloWorld")
}
}

