% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_requeststatus.R
\name{collection_requeststatus}
\alias{collection_requeststatus}
\title{Get request status}
\usage{
collection_requeststatus(conn, requestid, raw = FALSE, ...)
}
\arguments{
\item{conn}{A solrium connection object, see \link{SolrClient}}

\item{requestid}{(character) Required. The user defined request-id for the
request. This can be used to track the status of the submitted asynchronous
task. \code{-1} is a special request id which is used to cleanup the stored
states for all of the already completed/failed tasks.}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{...}{You can pass in parameters like \code{property.name=value}    to set
core property name to value. See the section Defining core.properties for
details on supported properties and values.
(https://lucene.apache.org/solr/guide/8_2/defining-core-properties.html)}
}
\description{
Request the status of an already submitted Asynchronous
Collection API call. This call is also used to clear up the stored statuses.
}
