% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\alias{cores}
\title{List collections or cores}
\usage{
collections(...)

cores(...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A character vector
}
\description{
List collections or cores
}
\details{
Calls \code{\link{collection_list}} or \code{\link{core_status}} internally, 
and parses out names for you.
}
\examples{
\dontrun{
# connect
solr_connect(verbose = FALSE)

# list collections
collections()

# list cores
cores()

# curl options
library("httr")
collections(config = verbose())
}
}

