% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_reload.R
\name{core_reload}
\alias{core_reload}
\title{Reload a core}
\usage{
core_reload(name, raw = FALSE, callopts = list())
}
\arguments{
\item{name}{(character) The name of the core. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{callopts}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Reload a core
}
\examples{
\dontrun{
# start Solr with Schemaless mode via the schemaless eg: bin/solr start -e schemaless
# you can create a new core like: bin/solr create -c corename
# where <corename> is the name for your core - or creaate as below

# connect
solr_connect()

# Status of particular cores
core_reload("gettingstarted")
core_status("gettingstarted")
}
}

