% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_requeststatus.R
\name{core_requeststatus}
\alias{core_requeststatus}
\title{Request status of asynchronous CoreAdmin API call}
\usage{
core_requeststatus(requestid, raw = FALSE, callopts = list())
}
\arguments{
\item{requestid}{The name of one of the cores to be removed. Required}

\item{raw}{(logical) If \code{TRUE}, returns raw data}

\item{callopts}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Request status of asynchronous CoreAdmin API call
}
\examples{
\dontrun{
# start Solr with Schemaless mode via the schemaless eg: bin/solr start -e schemaless

# FIXME: not tested yet...
# solr_connect()
# core_requeststatus(requestid = 1)
}
}

