% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\alias{delete_by_id}
\alias{delete_by_query}
\title{Delete documents by ID or query}
\usage{
delete_by_id(ids, commit = TRUE, commit_within = NULL, overwrite = TRUE,
  boost = NULL, wt = "json", raw = FALSE,
  base = "http://localhost:8983", ...)

delete_by_query(query, commit = TRUE, commit_within = NULL,
  overwrite = TRUE, boost = NULL, wt = "json", raw = FALSE,
  base = "http://localhost:8983", ...)
}
\arguments{
\item{ids}{Document IDs, one or more in a vector or list}

\item{commit}{(logical) If \code{TRUE}, documents immediately searchable.
Deafult: \code{TRUE}}

\item{commit_within}{(numeric) Milliseconds to commit the change, the document will be added
within that time. Default: NULL}

\item{overwrite}{(logical) Overwrite documents with matching keys. Default: \code{TRUE}}

\item{boost}{(numeric) Boost factor. Default: NULL}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to
parse}

\item{raw}{(logical) If \code{TRUE}, returns raw data in format specified by
\code{wt} param}

\item{base}{(character) URL endpoint. This is different from the other functions in
that we aren't hitting a search endpoint. Pass in here}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}

\item{query}{Query to use to delete documents}
}
\description{
Delete documents by ID or query
}
\examples{
\dontrun{
# Delete by ID
# delete_by_id(ids = 9)
## Many IDs
# delete_by_id(ids = c(3, 4))

# Delete by query
# delete_by_query(query = "manu:bank")
}
}

