% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{makemultiargs}
\alias{makemultiargs}
\title{Function to make make multiple args of the same name from a 
single input with length > 1}
\usage{
makemultiargs(x)
}
\arguments{
\item{x}{Value}
}
\description{
Function to make make multiple args of the same name from a 
single input with length > 1
}

