% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping a Solr instance}
\usage{
ping(conn, wt = "json", verbose = TRUE, raw = FALSE, callopts = list())
}
\arguments{
\item{conn}{Connection object. Required.}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to parse}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{raw}{(logical) If TRUE, returns raw data in format specified by wt param}

\item{callopts}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Ping a Solr instance
}
\details{
You likely may not be able to run this function against many public Solr
services, but should work locally.
}
\examples{
\dontrun{
# by default we connect to localhost, port 8983
conn <- solr_connect()
ping(conn)
ping(conn, wt = "xml")
ping(conn, verbose = FALSE)
ping(conn, raw = TRUE)

library("httr")
ping(conn, wt="xml", callopts = verbose())
}
}

