% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Ping a Solr instance}
\usage{
ping(base = "http://localhost:8983", wt = "json", verbose = TRUE,
  raw = FALSE, callopts = list())
}
\arguments{
\item{base}{(character) URL endpoint. This is different from the other functions in that we aren't
hitting a search endpoint. Pass in here}

\item{wt}{(character) One of json (default) or xml. If json, uses
\code{\link[jsonlite]{fromJSON}} to parse. If xml, uses \code{\link[XML]{xmlParse}} to parse}

\item{verbose}{If TRUE (default) the url call used printed to console.}

\item{raw}{(logical) If TRUE, returns raw data in format specified by wt param}

\item{callopts}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Ping a Solr instance
}
\details{
You likely may not be able to run this function against many public Solr
services, but should work locally.
}
\examples{
\dontrun{
ping()
ping('http://localhost:8983')
ping('http://localhost:8983', wt="xml")
ping('http://localhost:8983', verbose=FALSE)
ping('http://localhost:8983', raw=TRUE)

library("httr")
ping('http://localhost:8983', wt="xml", callopts = verbose())
}
}

