% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parsers.r
\name{pivot_flatten_tabular}
\alias{pivot_flatten_tabular}
\title{Flatten facet.pivot responses}
\usage{
pivot_flatten_tabular(df_w_pivot)
}
\arguments{
\item{df_w_pivot}{a \code{tbl_df} or \code{data.frame} with
nested \code{data.frame} objects representing nested
facet.pivot results}
}
\value{
a \code{tbl_df}
}
\description{
Convert a nested hierarchy of facet.pivot elements
to tabular data (rows and columns)
}
\keyword{internal}

