% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.r
\name{pivot_flatten_tabular}
\alias{pivot_flatten_tabular}
\title{Flatten facet.pivot responses}
\usage{
pivot_flatten_tabular(df_w_pivot)
}
\arguments{
\item{df_w_pivot}{a \code{data.frame} with another
\code{data.frame} nested inside representing a
pivot reponse}
}
\value{
a \code{data.frame}
}
\description{
Convert a nested hierarchy of facet.pivot elements
to tabular data (rows and columns)
}
\keyword{internal}
